Patch to requirements for legacy Genomics install
Globus genomics needs this stale, crufty version of the nexus client to be
installable. We should take the time at some point to fix that, but in the
meantime we need a working branch whose SHAs can be used for deployments
without having to manually patch requirements or do other hijinks. For now, the
only known fix that is needed is an update to the required version of httplib2